+++ include/transform.h
@@ -3,22 +3,22 @@
 #define TRANSFORM_H
 
 #include "linAlg.h"
-#include "Eigen/Geometry"
+#include "lib/Eigen/Geometry"
 #include <cmath>
 
 // returns 4x4 identity matrix
-static Matrix4f identity4f()
+static Matrix4d identity4f()
 {
-    Matrix4f out; 
+    Matrix4d out; 
     out.setIdentity();
     return out;
 }
 
 // returns 4x4 rotation matrix given axis-angle rotation
-static Matrix4f rotate(float x, float y, float z, float theta)
+static Matrix4d rotate(double x, double y, double z, double theta)
 {
-    Matrix4f m;
-    Vector3f axis(x, y, z);
+    Matrix4d m;
+    Vector3d axis(x, y, z);
 
     axis.normalize();
     // convert theta to radians
@@ -26,15 +26,15 @@ static Matrix4f rotate(float x, float y, float z, float theta)
    
     // make homogenous rotation matrix
     m.setIdentity();
-    m.block<3,3>(0,0) = (Matrix4f)Eigen::AngleAxisf(theta, axis);
+    m.block<3,3>(0,0) = (Matrix4d)Eigen::AngleAxisf(theta, axis);
 
     return m;
 }
 
 // returns 4x4 translation matrix
-static Matrix4f translate(float x, float y, float z)
+static Matrix4d translate(double x, double y, double z)
 {
-    Matrix4f m;
+    Matrix4d m;
     m.setIdentity();
     m(0,3) = x;
     m(1,3) = y;
@@ -44,9 +44,9 @@ static Matrix4f translate(float x, float y, float z)
 }
 
 // returns 4x4 scaling matrix
-static Matrix4f scale(float x, float y, float z)
+static Matrix4d scale(double x, double y, double z)
 {
-    Matrix4f m;
+    Matrix4d m;
     m << x, 0, 0, 0,
          0, y, 0, 0,
          0, 0, z, 0,
